src/requirements.txt
aniso8601==10.0.0
attrs==25.3.0
bcrypt==3.2.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
flasgger==0.5.4
Flask==3.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
marshmallow==4.0.0
passlib==1.7.4
pycparser==2.22
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
six==1.17.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3
